created useAppFocusEffect
