+++ backend/src/FitnessAppServer.ts
@@ -63,7 +63,7 @@ export class FitnessAppServer {
     );
 
     // const asdf = new MessageResolver();
-    // asdf.sendMessage(3, "ooga booga 15", {
+    // asdf.sendMessage(3, "ooga booga 16", {
     //   req: {
     //     session: { userId: 1 },
     //   },
